Fix infinite spinner on PIN login
Due to odd timing on PIN login, we drop onKeyListChange() callbacks since they come before we are logged in.
Hack a dispatch to updateWalletsRequest() which effectively enacts an onKeyListChange() immediately after login so that we get an updated list of wallets.

Long term, we need to fix the API to allow the GUI make a call when it is ready to receive callbacks after onLogin is received.

Code authored and reviewed by Kevin Sullivan