Fix feedback loop
Only set the overridePrimaryExchangeAmount in componentDidMount not in componentWillReceiveProps so that it is only overridden once and not when user types.
To work, this requires that a synchronous action updateParsedUri is exposed for Scan.ui.js to call before transitioning scenes to SendConfirmation. It needs redux state which contains the parsedUri to be latched before the scene change.