Fix the crash on sort with loading wallets
The basic idea is to stop hacking the GuiWallet objects we pull out of redux, but instead to simply pass walletId's to the two list views. Then, they can pull the pristine wallets out of Redux and pass them down to the rows. If there *is* no GuiWalllet, because it's still loading, we render a placeholder row instead.

This simplifies the props the row components expect, since they don't need to deal with missing wallets.