Eliminate WalletListStyles
The styles now live in each individual component. This creates some duplication, but we if we want to fix that, we should extract some common components. Sharing styles like this means we are likely to screw up one component while editing another.

Also, doing this leaves us with less code overall, since it's now possible to find & eliminate unused styles on a file-by-file basis. So, for all we gained in duplication, we lost in terms of dead code removal.