Remove unused & redundant theme colors
We use `colors.primary` for some modal circle icons and `THEME.COLORS.SECONDARY` for others, but visually they are the same color. Actually, `THEME.COLORS.SECONDARY` is the color of the ring around the icon, so it's even better, to match that.

The `transparent` CSS color is an alias for `rgba(0, 0, 0, 0)`, so just use that.