Fix crash when transaction is received before settings are loaded
If the core sends back a transaction before the front end has loaded settings, it's possible for us to have a blank entry for the denomination of a currency code. In such a case, return a blank currency code which will fall back to a dummy denomination object preventing a crash.