Fix falsy evaluation
uniqueIdentifiers can be an empty string so we need to use it as such if it is present. Since we're grabbing it from newSpendInfo (or state) already we shouldn't stomp on top of it with the result of getSpendTargetOtherParams.

(cherry picked from commit 6d587c6591f2baee25769da290c1b9a496e929c1)