Fix missing FIO handle on SendConfirmation
After choosing a FIO handle, handleChangeAddress must update the guiMakeSpendInfo through sendConfirmationUpdateTx. However, since there is no amount at the time the address is chosen, sendConfirmationUpdateTx early exits and the fioAddress in the guiMakeSpendInfo is lost.

Fix by dispatching an update with a null tx but with guiMakeSpendInfo

Also fix bug where SendScene connector tries to access this.props instead of 'props'