Update WalletListModal search state logic
The onBlur callback of OutlinedTextInput search box was hooked up to a call to setSearching(false), clearing the search whenever the user unfocused (onblur) the search box.

Additionally, the expected behavior is such that any time the search is focused OR the user has an active search entered (focused or not), the wallet list is NOT grouped.

Any time the  search input is blank AND unfocused, the wallet list IS grouped.

Update the search state logic such that the behavior is correct for all scenarios.