Add handling for undefined allowedCurrencyCodes from EdgeProviders
Changes to the wallet list asset filtering cause wallet lists shown for providers to crash when the providers provide invalid allowed currency codes.

The pre-existing functionality in these scenarios was to simply show a wallet list modal with no filtering.

Update so the pre-existing functionality is retained and all wallets are shown when invalid allowedCurrencyCodes are given from our providers.