Don't bother rounding yesterday's date
We can use the date as-is and let the server normalize. This change uncovered the (small and infrequent) possibility in useTokenDisplayData where the calculated yesterday date string might not exist in the exchange rate cache so the check is more general and just looks for a key with a date. We could get fancier with an actual date string regex but the extra underscore check is sufficient for how the cache is filled up and used.