Automate env.json setup during yarn prepare
The automation process aims to reduce the manual setup process for
`env.json`.

If `env.json` does not exist in the root project directory, then it will
be created with content mirroring that of `env.example.json`.

Otherwise, `env.json` will be updated to reflect the merge of `env.json`
and `env.example.json`. `config.js` recognizes `env.json` as the
canonical file where existing values will remain as-is, if the example
json file has empty corresponding values.

If new keys/fields are introduced in the example json file, these new
fields will be reflected in `env.json`.