Fix user cannot access account
App may crash when there is a race condition while trying to compute the
total balance of the account. The race condition is triggered by core
wallets and GUI wallets try to update at the same time.

Resolved the conflict by using the core wallets as a single source of
truth.