Fix CurrencyRow's CurrencyCode display
CurrencyRow is setup to accept both a token and tokenId. Should consider consolidating this in the future. Currently, it does not correctly set the currency code when setting only the tokenId and defaults to the wallet's native currency code.

Fix CurrencyRow so it can correctly show the currency code when only a tokenId is provided.