Move mockMode data field from ActionProgramState to ActionProgram
The motivation for persisting the mockMode field is to allow a dev to
toggle ENV.ACTION_QUEUE.mockMode without impacting running programs.
By placing the mockMode field in the ActionProgram, we can almost
guarantee this setting will not be changed because the ActionProgram
state is specified as be immutable.
If we leave it in the ActionProgramState, we risk it be unset from
a bug when returning `nextState` (unlikely, but still).