Fix race-ahead bug in action queue runtime
This is a regression from a previous refactor (AAVE: ActionQueue -
Implement Cancel Support).

The problem is that the index of completed actions always increments by
one when the effect is partially complete (a SeqEffect with childEffects
that have an index higher than opIndex). This is completely broken and
causes the SeqEffect to quickly pass through all partially complete
phases and ending at waiting on an non-effective effect to become
effect (last effect in the sequence, essentially)

To fix this bug, the partially completed effect handling must be dealt
with correctly.