Initialize numSelected to number of selected items
This is a potential fix for an issue where navigation back and forth through the scenes causes the next button to fail to appear sometimes even when items are selected. I wasn't able to reproduce, but if for some reason the numSelected state value is reinitialized then this way it reset to the correct number. That's assuming it's actually related to useState and not some flatlist display issue (like the selected state values actually did reset but the flatlist row visuals got stuck).