Call withServices HOC only once by moving invocation to `LoginScene.tsx`
It's a footgun to invoke an HOC within a component renders (Main)
because the HOC can (and usually does) return a new component function
on each invocation causing the render to always unmount the previous
component and mount the new component returned. This is a cause for
some really bad behavior.