Rewrite useSpaceStyles to only deal with "space props"
No more alignment props. In order to simulate alignment, you assign
a boolean to the space props. For example, to align left, you give the
component right space: `right={true}`. This is a leads to a very easy
mental model for using the component and a very minimal and elegant API.