Implement optimistic UI for exchange setting switch
With optimistic UI, the switch immediately reflects the change while
the request for the change is pending. This is better perceived-
performance to the user. 

This change is per Matthew's request. We can get away with this UX
pattern now that the scene is immediately issuing the request and 
listening to the core's state changes, eliminating a race condition to
show the user an "old" state.