Add init, approve, and reject session to wallet connect hook.
The functionality has changed a bit in this commit. Previously, we would need to grab a random wallet to parse the uri and did all the real work on the connect scene. Instead, we can initialize the connection on the connections scene before navigating away in case there is an error. Only if the init is successful do we proceed and show a user a list of wallets that are compatible with the session.

This change in behavior means we need to route to a different scene when deeplinking.

Also, dapps can provide svgs as logos but I hacked it to just use a wallet connect logo since I couldn't get the styling right in time.