Prevent hanging on a failed deleteSession call
This method is pretty unreliable. Internally, the sdk waits for the dapp to acknowledge the session and then emits the deletesession event. In testing, sometimes the dapp didn't respond or we did successfully send the delete session but some other internal issue in the sdk kept hanging indefinitely.

If the sdk takes too long emitting two messages and deleteing a local record then we just do it manually and hide it from the user.