Allow pasting into the text input again
The "suffix" feature on the FIO registration scene works by making the text input tiny. As the user enters text, the input gets wider, and the suffix moves to the right.

This means that the text input itself is tiny to start, so there is no tappable area for a user to long-press on. This is unacceptable, so the FIO effect needs to go.