Open links in a browser if account isn't logged in
We still want to avoid modals on top of modals as in the old implementation, so opening a browser window is the next best thing.

We may want to consider opening a browser for this modal in all cases (logged in/out) since the pages are difficult to navigate without a browser back button.