Fix estimateUnstakeFee for small unstake amounts
Prior logic is to pick the largest stake position in the pool to use as an example for how much slippage would occur when unstaking the requested amount.

However a very large position would make unstaking a very small amount unfeasable since unstakes are specified by bps.

Instead, search for the first position that is no larger than 10x the requested unstake amount