Add ability to reset SceneScrollState
Some scenes don't register the scroll handler and therefore don't have
local scroll state to stomp on the global scroll state. In this case,
they'll be left with another scene's scroll state.

To fix this, we set the scroll state back to its default scroll state
when the scene is no longer in focus.

We also fix a bug where local state never stomps on scrollXDelta
and scrollYDelta.