Convert footerHeight into a SceneWrapper prop
The only real reason to have footerHeight be global/shared state is to
use it in the `scrollDeltaToRatioDeltaFactor` calculation for the
gesture animation. We can trade-off this detail in the design for
performance by making the factor a constant and removing the global
state.

The other places where footerHeight is used is as layout values for the
SceneWrapper and it's children. So, by making it a prop, we can control
re-renders at the SceneWrapper call-site.