Remove tab height from insets when keyboard is open
This fixes having too much space when the keyboard is up which impacts
too much over-scroll on scroll components and cut-off on none-scroll
components. This only impacts scenes with MenuTabs, so as of committing,
that means the WalletListScene and TransactionListScenes were incorrect
because they include the tab height in their scroll view content
padding. They are now fixed with after this change.