Fix `keyboardType="default"` for Android Fix for the original reported issue. Unable to reproduce, so this is an educated shot in the dark.
This is not an exhaustive fix, but a targetted one limited in scope to the reported issue of the private seed modal's password input field being affected by this Android focus bug.

Technically `keyboardType="email-address"` is also bugged, but no reports yet of problems there, so hopefully we can retain "email-address" functionality.