Switch to search wallet list when search terms have been entered
Before this change, we render the search list during focus which caused
autofocus to effectively enable search mode in the WalletListModal to
be enabled as soon as the modal opens. Instead, we want to allow for
focus without render the search list, so that way you can still view
the SectionList version of the wallet list.