Fix the clipboard crash
The `ShareButtons` do not take any parameters on their callbacks, according to the types. That's not quite true, though, since the `ShareButtons` hand their callbacks on to `onPress`, and `onPress` provides the native event as a method parameter. The `copyToClipboard` method mistook the native touch event as its optional `uri` parameter, and then tried passing that to the clipboard. The clipboard library did NOT like this, and crashed violently.