Add `use_modular_headers!` to the Podfile
See the CocoaPods 1.5.0 release notes for context: https://blog.cocoapods.org/CocoaPods-1.5.0/

With this change, we no longer need to wac-a-mole individual places where a Swift pod depends on an Objective-C pod. Instead, we opt all libraries into using the stricter Objective-C import paths. There are no build breaks with this turned on, so it seems like all the pods we depend are already using header paths correctly.