Clean up push-message handling
It makes no sense to parse the message, convert it to a string, parse it again, and then handle it. Just handle it directly.

Also, move the price-change link type to the correct file. These `DeepLink` link types repsesent things the app can do. It doesn't matter where they come from, or how they are encoded.

Finally remove the "edge://price-change" link parsing, which isn't something we ever officially supported. This was an internal detail that leaked out as part of the silly parse / stringify / parse flow.