Fix display bug caused by flipping inputs on SwapCreateScene
Because we're flipping the wallets and the amounts for each SwapInput
component and each strategy to change this state is mixed between
imperative and declarative, we converge with incorrect state. This is
because the imperative change to the amounts happens to the SwapInput
before it's wallet state is change (which is the declarative state
change). Ultimately the mix of two techniques on state management is
the cause of the bug.

The solution is to add a delay to the imperative change, to allow for
the component's state to be updated before making the imperative
changes.