Add type annotatons to router components
With this change, we now verify that the scenes inside a router match what the router expects. This also gives us type checking for when we navigate to a scene nested within a router.

Unfortunately, we still have the `NavigationBase` type in the app which doesn't properly distinguish between the different types of navigation objects and their varying capabilities. This needs to be fixed.