Normalize error tracking
1. Isolation normalization into new utility to de-dup error handling
2. Use `normalizeError` util at the end (trackError)
3. Fully remove i18n from error objects; only present localized strings
   in the UI.

Translating error messages is be a locale issue. Maintaining our errors
in the default language is more consistent and consistency is better
when it comes to error tracking.

If "to translate" an error means something other than i18n, such as to
translate the error object and any special properties into a useful
message for reporting, then a new function should be introduced for
that purpose (call it `encodeError` or something of that nature).