Fix incorrect rounded corners on iOS
Regressed as a result of the latest `NotificationCard` updates (8d97172)

The above commit made changes to support rounded borders for the
`NotificationCard` by changing `BlurBackground` for all callers. The
callers responsible for the header/footer also use `BlurBackground,` but
require a non-rounded variant.

Due to the parity difference in iOS and Android blur styling, this
regression was only observable on iOS.