Make text input components semi-controlled
The reason for this change is to fix bugs on Android related to
controlled TextInput components. Cursor placement is not accurate when
entering text for a controlled TextInput on Android.

A semi-controlled component is one where the value of the input is
controlled by the component imperatively; not declaratively. This is
done via the `setNativeProps` method on the TextInput.

This is a nice solution because it abstracts away the imperative nature
of an uncontrolled component (TextInput) with a parent controlled
component (SimpleTextInput/FilledTextInput).