Properly handle min/max errors in Paybis ramp plugin
The implementation for this was missing from the paybisProvider, and
in addition to this missing implementation, there appears to be new
formats to min/max errors from the Paybis API, so we should account
for both the legacy errors observed to be a min/max error and the new
error messages for the min/max error messages. This detection is pretty
thorough.