Loosen the strictness of the vault records persisted
This relax on whether existing records follow the valid shape defined
by the cleaners is only because during development, the underlying
types change and cause a permanently broken state for the account.

In the future, we can harden this API to throw an error to not risk
losing data if we desire to.